pyproject.toml
[project]
name = "oni-utility-suite"
version = "0.1.0"
description = ""
authors = [
    {name = "KutologoJM",email = "kutologomashego@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-extensions (>=4.1,<5.0)",
    "pytest (>=8.4.2,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
